+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 Version 5.2-SNAPSHOT
 -------------
 
+ADDED:
+
+- new method for get the default field names that can be always ignored that can be used in unit tests
+
+CHANGED:
+
+- bugfix for method ReflectionExtensions#getAllDeclaredFields(Class, String...) for find all declared fields
+
 Version 5.1
 -------------